fix(types): fix quick win type errors across 3 files (6→0 errors, 100%)
**drug_resistance_analyzer.py (1→0):**
- Line 561: Add explicit float() cast to fix no-any-return error
- Ensures _calculate_hotspot_risk_score returns proper float type

**treatment_protocol_engine.py (1→0):**
- Line 582: Replace None return with ValueError exception
- _get_default_who_protocol now properly validates protocols are loaded

**monitoring_dashboards.py (4→0):**
- Add cast() imports for proper type narrowing
- Line 23: Cast json.load() result to dict[str, Any]
- Lines 30, 35: Cast config.get() results to list[dict[str, Any]]
- Line 41: Add explicit type annotation for rules dict

+++ src/malaria_predictor/healthcare/treatment_protocols/drug_resistance_analyzer.py
@@ -558,7 +558,7 @@ class DrugResistanceAnalyzer:
             "low": 0.6
         }.get(location_data.get("surveillance_quality", "medium"), 0.8)
 
-        return base_score * population_factor * quality_multiplier
+        return float(base_score * population_factor * quality_multiplier)
 
     # Additional helper methods would be implemented here
     def _get_resistance_profile(self, drug: str, location: str) -> dict:
